Entite

+++ src/main/java/universite_paris8/iut/mcheema/codesource/modele/batiment/Cloud.java
@@ -5,7 +5,7 @@ import universite_paris8.iut.mcheema.codesource.modele.ennemi.Ennemi;
 import universite_paris8.iut.mcheema.codesource.modele.projectile.Projectile;
 
 public class Cloud extends BatimentTir {
-    public Cloud(int x, int y, Environnement env) {
+    public Cloud(double x, double y, Environnement env) {
         super("Cloud", x, y, 50, 1, 50, 100, env);
     }
